+++ nservicebus/azure-service-bus/sanitization.md
@@ -23,15 +23,15 @@ When entity name does not follow the rules, an exception is raised by the broker
 
 | Entity Type  | Valid Characters | Path / Name Maximum Length |
 |--------------|------------------|----------------------------|
-| Queue        | Letters, numbers, periods (.), hyphens (-), underscores (_), and forward slashes (/). | 260 |
-| Topic        | Letters, numbers, periods (.), hyphens (-), underscores (_), and forward slashes (/). | 260 |
-| Subscription | Letters, numbers, periods (.), hyphens (-), and underscores (_). | 50  |
-| Rule         | Letters, numbers, periods (.), hyphens (-), and underscores (_). | 50  |
+| Queue        | Letters, numbers, periods (`.`), hyphens (`-`), underscores (`_`), and forward slashes (/). | 260 |
+| Topic        | Letters, numbers, periods (`.`), hyphens (`-`), underscores (`_`), and forward slashes (/). | 260 |
+| Subscription | Letters, numbers, periods (`.`), hyphens (`-`), and underscores (`_`). | 50  |
+| Rule         | Letters, numbers, periods (`.`), hyphens (`-`), and underscores (`_`). | 50  |
 
 
 ## Sanitization
 
-Sanitization is an operation on entity path or name to ensure the broker can operate on the entity, and no exception is thrown. Sanitization can be performed manually or by the Azure Service Bus transport.
+Sanitization is an operation that is performed on an entity path or name to ensure the broker can operate on the entity, and no exception is thrown. Sanitization can be performed manually or by the Azure Service Bus transport.
 
 To perform manual sanitization, inspect entity name for invalid characters and number of characters. For automated sanitization, the transport allows configuration outlined below.
 
@@ -41,7 +41,7 @@ To perform manual sanitization, inspect entity name for invalid characters and n
 
 ### Versions 6 and below
 
-All entities treated the same for allowed characters. Only letters, numbers, periods (.), hyphens (-), and underscores (_) were allowed. Maximum length was predefined by the transport.
+All entities treated the same for allowed characters. Only letters, numbers, periods (`.`), hyphens (`-`), and underscores (`_`) were allowed. Maximum length was predefined by the transport.
 
 In Version 6.4.x [Naming Conventions](/nservicebus/azure-service-bus/native-integration.md) were added to allow custom sanitization of queues, topics, and subscriptions.
 
@@ -58,14 +58,14 @@ Custom sanitization strategy definition:
 
 snippet: asb-custom-sanitization-strategy
 
-When using `EndpointOrientedTopology`, the transport should be configured to use `EndpointOrientedTopologySanitization` sanitization strategy for backward compatibility with endpoints version 6 and below.
+When using `EndpointOrientedTopology`, the transport should be configured to use `EndpointOrientedTopologySanitization` sanitization strategy for backward compatibility with endpoints versions 6 and below.
 
 snippet: asb-endpointorientedtopology-sanitization
 
 
 ## Future consideration prior to using sanitization
 
-When implementing a custom sanitization, consider factors such as readability and discover-ability. Things to consider:
+When implementing custom sanitization, consider factors such as readability and discover-ability. Things to consider:
 
  * Truncated long entity names could conflict. 
  * Hashed entity names could lead to difficult names to use during production troubleshooting or debugging.
