+++ samples/azure/azure-service-bus/ASB_7/Endpoint1/app.config
@@ -46,8 +46,4 @@
 </bindingExtensions>
 </extensions>
 </system.serviceModel>
-<appSettings>
-<!-- Service Bus specific app setings for messaging connections -->
-<add key="Microsoft.ServiceBus.ConnectionString" value="Endpoint=sb://[your namespace].servicebus.windows.net;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=[your secret]" />
-</appSettings>
 </configuration>
