+++ samples/azure/azure-service-bus/ASB_7/Shared/app.config
@@ -42,8 +42,4 @@
       </bindingExtensions>
     </extensions>
   </system.serviceModel>
-  <appSettings>
-    <!-- Service Bus specific app setings for messaging connections -->
-    <add key="Microsoft.ServiceBus.ConnectionString" value="Endpoint=sb://[your namespace].servicebus.windows.net;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=[your secret]" />
-  </appSettings>
 </configuration>
