+++ samples/azure/custom-sanitization/ASB_7/Subscriber/Program.cs
@@ -1,12 +1,7 @@
 ﻿using System;
-using System.Linq;
-using System.Security.Cryptography;
-using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using NServiceBus;
 using NServiceBus.AzureServiceBus;
-using NServiceBus.AzureServiceBus.Addressing;
 using NServiceBus.Features;
 
 class Program
@@ -51,74 +46,3 @@ class Program
         }
     }
 }
-
-#region Sha1SanitiazationStrategy
-
-class Sha1Sanitization : ISanitizationStrategy
-{
-    public string Sanitize(string entityPathOrName, EntityType entityType)
-    {
-        // remove invalid characters
-        if (entityType == EntityType.Queue || entityType == EntityType.Topic)
-        {
-            var regexQueueAndTopicValidCharacters = new Regex(@"[^a-zA-Z0-9\-\._\/]");
-            var regexLeadingAndTrailingForwardSlashes = new Regex(@"^\/|\/$");
-
-            entityPathOrName = regexQueueAndTopicValidCharacters.Replace(entityPathOrName, string.Empty);
-            entityPathOrName = regexLeadingAndTrailingForwardSlashes.Replace(entityPathOrName, string.Empty);
-        }
-
-        if (entityType == EntityType.Subscription || entityType == EntityType.Rule || entityType == EntityType.EventHub)
-        {
-            var rgx = new Regex(@"[^a-zA-Z0-9\-\._]");
-            entityPathOrName = rgx.Replace(entityPathOrName, "");
-        }
-
-        int entityPathOrNameMaxLength = 0;
-
-        switch (entityType)
-        {
-            case EntityType.Queue:
-            case EntityType.Topic:
-                entityPathOrNameMaxLength = 260;
-                break;
-            case EntityType.Subscription:
-            case EntityType.Rule:
-                entityPathOrNameMaxLength = 50;
-                break;
-        }
-
-        // hash if too long
-        if (entityPathOrName.Length > entityPathOrNameMaxLength)
-        {
-            entityPathOrName = SHA1DeterministicNameBuilder.Build(entityPathOrName);
-        }
-
-        return entityPathOrName;
-    }
-}
-
-#endregion
-
-#region SHA1DeterministicNameBuilder
-
-static class SHA1DeterministicNameBuilder
-{
-    public static string Build(string input)
-    {
-        using (var provider = new SHA1CryptoServiceProvider())
-        {
-            var inputBytes = Encoding.Default.GetBytes(input);
-            var hashBytes = provider.ComputeHash(inputBytes);
-
-            var hashBuilder = new StringBuilder(string.Join("", hashBytes.Select(x => x.ToString("x2"))));
-            foreach (var delimeterIndex in new[] { 5, 11, 17, 23, 29, 35, 41 })
-            {
-                hashBuilder.Insert(delimeterIndex, "-");
-            }
-            return hashBuilder.ToString();
-        }
-    }
-}
-
-#endregion
