+++ samples/azure/custom-sanitization/ASB_7/Subscriber/app.config
@@ -41,42 +41,23 @@
   </appSettings>
 
   <runtime>
-
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-
       <dependentAssembly>
-
         <assemblyIdentity name="Microsoft.Data.Edm" publicKeyToken="31bf3856ad364e35" culture="neutral" />
-
         <bindingRedirect oldVersion="0.0.0.0-5.7.0.0" newVersion="5.7.0.0" />
-
       </dependentAssembly>
-
       <dependentAssembly>
-
         <assemblyIdentity name="Microsoft.Data.Services.Client" publicKeyToken="31bf3856ad364e35" culture="neutral" />
-
         <bindingRedirect oldVersion="0.0.0.0-5.7.0.0" newVersion="5.7.0.0" />
-
       </dependentAssembly>
-
       <dependentAssembly>
-
         <assemblyIdentity name="Microsoft.Data.OData" publicKeyToken="31bf3856ad364e35" culture="neutral" />
-
         <bindingRedirect oldVersion="0.0.0.0-5.7.0.0" newVersion="5.7.0.0" />
-
       </dependentAssembly>
-
       <dependentAssembly>
-
         <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
-
         <bindingRedirect oldVersion="0.0.0.0-8.0.0.0" newVersion="8.0.0.0" />
-
       </dependentAssembly>
-
     </assemblyBinding>
-
   </runtime>
 </configuration>
